User Account state
User accounts now carry active state.  If an admin user creates an account the state is set to active and the user can login.  I a person creates a new account from the signup page the active state is set to false.  An admin can approve the account.

We used to not ign a role to a new user, therefore denying them access to any data.  Users could login without roles they just could not see anything.

We will now require all new users have a 'default' role, for now that is USER_ROLE.

fixes #27