Image processing in a queue
Changes to how attachments are processed server side.  Images will now be moved to their home on the file system right away and the server will return with the appropriate response right away.  All processing will be added to a queue (size is in configuration), i.e. orientation and thumb-nailing.  Since these image operations could be expensive and might take some time, esp. on a tiny server we want to limit the amount of images that can be processed simultaneously.