New observation popups
When you click on an observation on the map a leaflet popup is displayed.  This is instead of
drawing a purple circle around the observation and auto scorlling the feed.  There is an info
button in the pop-up that when clicked will scroll the feed to that observation.  This is
more consistient to how the mobile apps work.

Also clicks form the feed will pan (not zoom) the map to center around that observation and display the popup.