Fixed up py css that controlled padding to account for map
We were adding padding the wrong way to account for the navbar, which also made us pad all leaflet controls.
This meant that leaflet controls that were not under a navbar were padded when they should not have been padded.

Added new control to zoom to world extent.  This is used the the observation map clip to quickly zoom out to a world level.