Bootstrap 3+
Started to have problems with our old angular bootstrap plugins.  Unfortunately because we were on
Bootstrap 2.0 I was unable to upgrade to the newest angular bootstrap plugins as they dropped
support for bootstrap 2.0.

Finally took the plunge and upgraded to Bootstrap 3.+.

Also upgraded to latest bootstrap angular plugin.

Do I think Bootstrap 3.+ is the 'right' thing for MAGE web?  Nope.  However bootstrap 3 upgrade
was mostly search replace so it only took a day to upgrade and is a great stop gap for now.